Update fork to our needs.
Update package name so that we can publish it under our org.
Defined some working npm scripts to build the js and css.
Defined a prepublishOnly script to run all the build tasks before a new version is published.
Started to refactor some of the code to generate the necessary ARIA attributes.